delegate saving/restoring of chart state (layout, preference and drawings) into ChartState
