+++ src/store/StudyLegendStore.js
@@ -39,6 +39,14 @@ export default class StudyLegendStore {
         this.stx.callbacks.studyOverlayEdit = this.editStudy;
         this.stx.callbacks.studyPanelEdit = this.editStudy;
         this.stx.append('createDataSet', this.renderLegend);
+        this.stx.append('adjustPanelPositions', () => {
+            const panel = Object.keys(this.stx.panels)[1];
+            if (panel) {
+                // Hide the up arrow from first indicator to prevent user
+                // from moving the indicator panel above the main chart
+                this.stx.panels[panel].up.style.display = 'none';
+            }
+        });
         this.renderLegend();
     };
